+++ examples/plot_aia_overplotting.py
@@ -58,13 +58,15 @@ dtime = aia.date
 # at your own risk!
 [[br, header]] = sunpy.io.fits.read('190310t0014gong.fits')
 br = br - np.mean(br)
-###############################################################################
-# GONG maps have their LH edge at -180deg in Carrington Longitude,
-# so roll to get it at 0deg. This way the input magnetic field is in a
-# Carrington frame of reference, which matters later when lining the field
-# lines up with the AIA image.
-br = np.roll(br, header['CRVAL1'] + 180, axis=1)
 
+###############################################################################
+# GONG maps have their LH edge at varying Carrington longitudes,
+# so roll to get it at -180deg, which is what the carr_cea_wcs_header function
+# expects.
+br = np.roll(br, header['CRVAL1'], axis=1)
+
+header = pfsspy.carr_cea_wcs_header(aia.date, br.shape)
+gong_map = sunpy.map.Map((br, header))
 
 ###############################################################################
 # The PFSS solution is calculated on a regular 3D grid in (phi, s, rho), where
@@ -76,13 +78,15 @@ rss = 2.5
 ###############################################################################
 # From the boundary condition, number of radial grid points, and source
 # surface, we now construct an `Input` object that stores this information
-input = pfsspy.Input(br, nrho, rss, dtime=dtime)
+input = pfsspy.Input(gong_map, nrho, rss)
 
 ###############################################################################
 # Using the `Input` object, plot the input photospheric magnetic field
-fig, ax = plt.subplots()
-mesh = input.plot_input(ax)
-fig.colorbar(mesh)
+m = input.map
+fig = plt.figure()
+ax = plt.subplot(projection=m)
+m.plot()
+plt.colorbar()
 ax.set_title('Input field')
 
 ###############################################################################
@@ -105,13 +109,17 @@ s, phi = np.meshgrid(np.linspace(0.1, 0.2, 5),
 lat = np.arcsin(s) * u.rad
 lon = phi * u.rad
 
-fig, ax = plt.subplots()
-mesh = input.plot_input(ax)
-fig.colorbar(mesh)
-ax.scatter(np.rad2deg(phi), s, color='k', s=1)
+m = input.map
+fig = plt.figure()
+ax = plt.subplot(projection=m)
+m.plot()
+plt.colorbar()
 
-ax.set_xlim(50, 70)
-ax.set_ylim(0, 0.35)
+ax.scatter(np.rad2deg(phi), np.rad2deg(np.arcsin(s)),
+           color='k', s=1, transform=ax.get_transform('world'))
+
+# ax.set_xlim(50, 70)
+# ax.set_ylim(0, 0.35)
 ax.set_title('Field line footpoints')
 
 #######################################################################
@@ -120,22 +128,25 @@ output = pfsspy.pfss(input)
 
 ###############################################################################
 # Trace field lines from the footpoints defined above.
-tracer = tracing.PythonTracer(atol=1e-6)
+tracer = tracing.PythonTracer()
 seeds = SkyCoord(lon.ravel(), lat.ravel(), 1.01 * const.R_sun, frame=output.coordinate_frame)
 flines = tracer.trace(seeds, output)
 
 ###############################################################################
 # Plot the input GONG magnetic field map, along with the traced mangetic field
 # lines.
-fig, ax = plt.subplots()
-mesh = input.plot_input(ax)
-for fline in flines:
-    fline.coords.representation_type = 'spherical'
-    ax.plot(fline.coords.lon / u.deg, np.sin(fline.coords.lat),
-            color='black', linewidth=1)
+m = input.map
+fig = plt.figure()
+ax = plt.subplot(projection=m)
+m.plot()
+plt.colorbar()
 
-ax.set_xlim(55, 65)
-ax.set_ylim(0.1, 0.25)
+for fline in flines:
+    coords = fline.coords.transform_to(m.coordinate_frame)
+    ax.plot_coord(coords, color='black', linewidth=1)
+
+# ax.set_xlim(55, 65)
+# ax.set_ylim(0.1, 0.25)
 ax.set_title('Photospheric field and traced field lines')
 ###############################################################################
 # Plot the AIA map, along with the traced magnetic field lines. Inside the