+++ pfsspy/tests/example_maps.py
@@ -1,5 +1,7 @@
 import pytest
 import numpy as np
+from sunpy.map import Map
+from astropy.time import Time
 import pfsspy
 
 
@@ -11,19 +13,18 @@ def zero_map():
     nr = 10
     rss = 2.5
     br = np.zeros((ns, nphi))
+    header = pfsspy.carr_cea_wcs_header(Time('1992-12-21'), br.shape)
+    input_map = Map((br, header))
 
-    input = pfsspy.Input(br, nr, rss)
+    input = pfsspy.Input(input_map, nr, rss)
     output = pfsspy.pfss(input)
     return input, output
 
 
 @pytest.fixture
 def dipole_map():
-    # Test a completely zero input
     ntheta = 30
     nphi = 20
-    nr = 10
-    rss = 2.5
 
     phi = np.linspace(0, 2 * np.pi, nphi)
     theta = np.linspace(-np.pi / 2, np.pi / 2, ntheta)
@@ -33,6 +34,15 @@ def dipole_map():
         return 2 * np.sin(theta) / r**3
 
     br = dipole_Br(1, theta).T
-    input = pfsspy.Input(br, nr, rss)
+    header = pfsspy.carr_cea_wcs_header(Time('1992-12-21'), br.shape)
+    return Map((br, header))
+
+
+@pytest.fixture
+def dipole_result(dipole_map):
+    nr = 10
+    rss = 2.5
+
+    input = pfsspy.Input(dipole_map, nr, rss)
     output = pfsspy.pfss(input)
     return input, output