+++ pfsspy/tests/test_pfss.py
@@ -3,22 +3,25 @@ import astropy.coordinates as coord
 import astropy.units as u
 from astropy.tests.helper import quantity_allclose
 
+import pytest
+
 import matplotlib
 import numpy as np
 import pfsspy
 import sunpy.map
+import sunpy.util.exceptions
 
 import pfsspy.coords
 from pfsspy import tracing
 
-from .example_maps import dipole_map, zero_map
+from .example_maps import dipole_map, zero_map, dipole_result
 matplotlib.use('Agg')
 
 R_sun = const.R_sun
 
 
-def test_expansion_factor(dipole_map):
-    inp, out = dipole_map
+def test_expansion_factor(dipole_result):
+    inp, out = dipole_result
     out_frame = out.coordinate_frame
 
     tracer = tracing.PythonTracer()
@@ -42,8 +45,8 @@ def test_expansion_factor(dipole_map):
     assert pil_field_line.expansion_factor > field_line.expansion_factor
 
 
-def test_field_line_polarity(dipole_map):
-    input, out = dipole_map
+def test_field_line_polarity(dipole_result):
+    input, out = dipole_result
     out_frame = out.coordinate_frame
 
     tracer = tracing.PythonTracer()
@@ -61,8 +64,8 @@ def test_field_line_polarity(dipole_map):
     assert eq_field_line.polarity == 0
 
 
-def test_footpoints(dipole_map):
-    input, out = dipole_map
+def test_footpoints(dipole_result):
+    input, out = dipole_result
     out_frame = out.coordinate_frame
 
     tracer = tracing.PythonTracer(atol=1e-8, rtol=1e-8)
@@ -118,39 +121,12 @@ def test_shape(zero_map):
     assert bg.shape == (nphi + 1, ns + 1, nr + 1, 3)
 
 
-def test_sunpy_map_input(zero_map):
-    zero_in, _ = zero_map
-    # Check that loading an input map works
-    header = {'cunit1': 'degree', 'cunit2': 'degree'}
-    map = sunpy.map.Map((zero_in.br, header))
-    input = pfsspy.Input(map, zero_in.grid.nr, zero_in.grid.rss)
-    assert (input.br == zero_in.br).all()
-
-
-def test_input_output(dipole_map):
+def test_input_output(dipole_result):
     # Smoke test of saving/loading files
-    _, out = dipole_map
+    _, out = dipole_result
     out.save('test.npz')
     new_out = pfsspy.load_output('test.npz')
     assert (new_out.al[0] == out.al[0]).all()
-
-
-def test_plot_input(dipole_map):
-    # Smoke test of input plotting
-    inp, out = dipole_map
-    inp.plot_input()
-
-
-def test_plot_source_surface(dipole_map):
-    # Smoke test of source surface plotting
-    inp, out = dipole_map
-    out.plot_source_surface()
-
-
-def test_plot_pil(dipole_map):
-    # Smoke test of PIL plotting
-    inp, out = dipole_map
-    out.plot_pil()
 
 
 def test_header_generation():
@@ -174,3 +150,14 @@ def test_header_generation():
     assert header['CRVAL2'] == 0
     assert header['CUNIT1'] == 'deg'
     assert header['CUNIT2'] == 'deg'
+
+
+def test_wrong_projection_error(dipole_map):
+    dipole_map.meta['ctype1'] = 'HGLN-CAR'
+    with pytest.raises(ValueError, match='must be CEA'):
+        pfsspy.Input(dipole_map, 5, 2.5)
+
+
+def test_non_map_input():
+    with pytest.raises(ValueError, match='br must be a SunPy Map'):
+        pfsspy.Input(np.random.rand(2, 2), 1, 1)